separate request building from response handling with `SessionTask` API
