+++ ELWebServiceExample/ELWebServiceExample/HTTP/BrewClientAPI.swift
@@ -12,22 +12,30 @@ import ELWebService
 protocol BrewClientAPI {
     var webService: WebService { get }
     
-    func fetchBrew(brewID: String) -> ServiceTask
-    func fetchAllBrews() -> ServiceTask
-    func save(brew: Brew) -> ServiceTask
+    func fetchBrew(brewID: String) -> SessionTask
+    func fetchAllBrews() -> SessionTask
+    func save(brew: Brew) -> SessionTask
 }
 
 /// Implements the web services for the HTTP client methods
 extension BrewClientAPI {
-    func fetchBrew(brewID: String) -> ServiceTask {
-        return webService.GET("/brews/\(brewID)")
+    func fetchBrew(brewID: String) -> SessionTask {
+        let request = webService.request(.GET, path: "/brews")
+        return sessionTask(request: request)
     }
     
-    func fetchAllBrews() -> ServiceTask {
-        return webService.GET("/brews")
+    func fetchAllBrews() -> SessionTask {
+        let request = webService.request(.GET, path: "/brews")
+        return sessionTask(request: request)
     }
     
-    func save(brew: Brew) -> ServiceTask {
-        return webService.POST("/brews").setFormParameters(brew.webServiceParameters)
+    func save(brew: Brew) -> SessionTask {
+        var request = webService.request(.POST, path: "/brews")
+        request.formParameters = brew.webServiceParameters
+        return webService.serviceTask(request: request)
+    }
+    
+    func sessionTask(request: URLRequestEncodable) -> SessionTask {
+        return webService.sessionTask(request: request)
     }
 }
